+++ test/rules/completed-docs/functions/overloads-separate-docs/false/tslint.json
@@ -2,7 +2,7 @@
   "rules": {
     "completed-docs": [true, {
       "functions": {
-        "overloads-separate-docs": false,
+        "overloads": false,
         "privacies": ["all"]
       }
     }]
