+++ test/rules/completed-docs/functions/overloads-separate-docs/true/tslint.json
@@ -2,7 +2,7 @@
   "rules": {
     "completed-docs": [true, {
       "functions": {
-        "overloads-separate-docs": true,
+        "overloads": true,
         "privacies": ["all"]
       }
     }]
