+++ test/rules/completed-docs/methods/overloads-separate-docs/false/tslint.json
@@ -2,7 +2,7 @@
   "rules": {
     "completed-docs": [true, {
       "methods": {
-        "overloads-separate-docs": false,
+        "overloads": false,
         "privacies": ["all"]
       }
     }]
