+++ test/rules/completed-docs/methods/overloads-separate-docs/true/tslint.json
@@ -2,7 +2,7 @@
   "rules": {
     "completed-docs": [true, {
       "methods": {
-        "overloads-separate-docs": true,
+        "overloads": true,
         "privacies": ["all"]
       }
     }]
